Re organize file structure
