+++ docs/source/rex/rex.multi_year_resource.rst
@@ -1,7 +1,7 @@
-rex.mutli\_year\_resource module
+rex.multi\_year\_resource module
 ================================
 
-.. automodule:: rex.mutli_year_resource
+.. automodule:: rex.multi_year_resource
    :members:
    :undoc-members:
    :show-inheritance:
