--- docs/source/rex/rex.multi_year_resource_cli.rst
@@ -1,11 +1,0 @@
-rex.multi\_year\_resource\_cli module
-=====================================
-
-.. click:: rex.multi_year_resource_cli:main
-   :prog: MultiYearX
-   :show-nested:
-
-.. automodule:: rex.multi_year_resource_cli
-   :members:
-   :undoc-members:
-   :show-inheritance:
