--- docs/source/rex/rex.resource_cli.rst
@@ -1,11 +1,0 @@
-rex.resource\_cli module
-========================
-
-.. click:: rex.resource_cli:main
-   :prog: rex
-   :show-nested:
-
-.. automodule:: rex.resource_cli
-   :members:
-   :undoc-members:
-   :show-inheritance:
