+++ docs/source/rex/rex.resource_extraction.multi_year_resource_cli.rst
@@ -1,0 +1,11 @@
+rex.resource\_extraction.multi\_year\_resource\_cli module
+==========================================================
+
+.. click:: rex.resource_extraction.multi_year_resource_cli:main
+   :prog: MultiYearX
+   :show-nested:
+
+.. automodule:: rex.resource_extraction.multi_year_resource_cli
+   :members:
+   :undoc-members:
+   :show-inheritance:
