+++ docs/source/rex/rex.resource_extraction.nsrdb_cli.rst
@@ -1,0 +1,11 @@
+rex.resource\_extraction.solar\_cli module
+==========================================
+
+.. click:: rex.resource_extraction.solar_cli:main
+   :prog: NSRDBX
+   :show-nested:
+
+.. automodule:: rex.resource_extraction.solar_cli
+   :members:
+   :undoc-members:
+   :show-inheritance:
