+++ docs/source/rex/rex.resource_extraction.resource_cli.rst
@@ -1,0 +1,11 @@
+rex.resource\_extraction.resource\_cli module
+=============================================
+
+.. click:: rex.resource_extraction.resource_cli:main
+   :prog: rex
+   :show-nested:
+
+.. automodule:: rex.resource_extraction.resource_cli
+   :members:
+   :undoc-members:
+   :show-inheritance:
