+++ docs/source/rex/rex.resource_extraction.resource_extraction.rst
@@ -1,0 +1,7 @@
+rex.resource\_extraction.resource\_extraction module
+====================================================
+
+.. automodule:: rex.resource_extraction.resource_extraction
+   :members:
+   :undoc-members:
+   :show-inheritance:
