+++ docs/source/rex/rex.resource_extraction.rst
@@ -1,7 +1,18 @@
-rex.resource\_extraction module
-===============================
+rex.resource\_extraction package
+================================
 
 .. automodule:: rex.resource_extraction
    :members:
    :undoc-members:
    :show-inheritance:
+
+Submodules
+----------
+
+.. toctree::
+
+   rex.resource_extraction.multi_year_resource_cli
+   rex.resource_extraction.resource_cli
+   rex.resource_extraction.resource_extraction
+   rex.resource_extraction.solar_cli
+   rex.resource_extraction.wind_cli
