+++ docs/source/rex/rex.resource_extraction.wind_cli.rst
@@ -1,0 +1,11 @@
+rex.resource\_extraction.wind\_cli module
+=========================================
+
+.. click:: rex.resource_extraction.wind_cli:main
+   :prog: WINDX
+   :show-nested:
+
+.. automodule:: rex.resource_extraction.wind_cli
+   :members:
+   :undoc-members:
+   :show-inheritance:
