--- docs/source/rex/rex.solar_cli.rst
@@ -1,11 +1,0 @@
-rex.solar\_cli module
-=====================
-
-.. click:: rex.solar_cli:main
-   :prog: NSRDB
-   :show-nested:
-
-.. automodule:: rex.solar_cli
-   :members:
-   :undoc-members:
-   :show-inheritance:
