--- docs/source/rex/rex.wind_cli.rst
@@ -1,11 +1,0 @@
-rex.wind\_cli module
-====================
-
-.. click:: rex.wind_cli:main
-   :prog: WIND
-   :show-nested:
-
-.. automodule:: rex.wind_cli
-   :members:
-   :undoc-members:
-   :show-inheritance:
