+++ rex/resource_extraction/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+"""
+Resource extractors
+"""
+from .resource_extraction import (ResourceX, MultiYearResourceX,
+                                  NSRDBX, MultiFileNSRDBX, MultiYearNSRDBX,
+                                  WindX, MultiFileWindX, MultiYearWindX)
